requirements.txt
# Core dependencies
requests
beautifulsoup4
python-nmap
selenium
pandas
openai
python-dotenv

# Reporting
Jinja2
matplotlib
pdfkit